Some cleanup (#23)
Cleaned up formatting with Columns and Rows (and Flexes) to try to make the indenting a bit more predictable. Rows and Columns can have a `content` size which will push out the margins so the Row/Column content aligns with outside content. Other sizes, or Rows/Columns with the `noSpace` attribute, will not push out.

As per the above, removed the `flush` attribute for Row.

Removed the growItems attribute on Flex, as in practice you need to apply `flex` styling to child elements anyway to get the behavior you want. Supporting one specific behavior via attributes isn't that useful.

Modified the alignment styles for Checkbox, Radio and Switch to get them to behave better if they're in a flex element with `stretch` applied. The label was aligned with the baseline but the checkbox/radio/switch was aligning to the middle of its container, which looked weird.

Renamed `styleguide` task to `start`, so it's consistent with our other projects.

Updated the Button and Switch demo to have interactive controls everywhere, instead of just the Toggle demo.